[FEATURE] v13 support
- Removed support for TYPO3 11 and 12
- Removed support for PHP7 and PHP8.0
- Added DDEV setup for development purposes
- Moved driver functionality to AbstractAmazonS3Driver to create AmazonS3Driver based on the existence of Capabilities class
- Updated Documentation
- Optimized code
