+++ .ddev/commands/host/docs
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker run --rm --pull always -v $(pwd):/project -it ghcr.io/typo3-documentation/render-guides:latest --config=Documentation && open $(pwd)/Documentation-GENERATED-temp/Index.html
