+++ .ddev/commands/host/minio
@@ -1,0 +1,17 @@
+#!/usr/bin/env bash
+
+## #ddev-generated: If you want to edit and own this file, remove this line.
+## Description: Launch a browser with Minio console
+## Usage: minio
+## Example: "ddev minio"
+
+echo "Login using 'ddevminio' user and 'ddevminio' password"
+
+DDEV_MINIOCONSOLE_PORT=9089
+DDEV_MINIOCONSOLE_HTTPS_PORT=9090
+
+if [ ${DDEV_PRIMARY_URL%://*} = "http" ] || [ -n "${GITPOD_WORKSPACE_ID:-}" ] || [ "${CODESPACES:-}" = "true" ]; then
+    ddev launch :$DDEV_MINIOCONSOLE_PORT
+else
+    ddev launch :$DDEV_MINIOCONSOLE_HTTPS_PORT
+fi
