+++ .ddev/commands/web/install-all
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+ABSOLUTE_PATH="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)/$(basename "${BASH_SOURCE[0]}/.")"
+
+$ABSOLUTE_PATH/install-v12
+$ABSOLUTE_PATH/install-v13
