+++ .ddev/docker-compose.docs.yaml
@@ -1,0 +1,7 @@
+services:
+    docs:
+        image: ghcr.io/t3docs/render-documentation:latest
+        command: makehtml
+        volumes:
+            - ../:/PROJECT:ro
+            - ../Documentation-GENERATED-temp:/RESULT
