+++ .ddev/docker-compose.web.yaml
@@ -1,0 +1,41 @@
+services:
+    web:
+        environment:
+            - EXTENSION_KEY=fal_s3
+            - PACKAGE_NAME=maxserv/fal_s3
+
+            # TYPO3 v12 config
+            - TYPO3_INSTALL_DB_DRIVER=mysqli
+            - TYPO3_INSTALL_DB_USER=root
+            - TYPO3_INSTALL_DB_PASSWORD=root
+            - TYPO3_INSTALL_DB_HOST=db
+            - TYPO3_INSTALL_DB_UNIX_SOCKET=
+            - TYPO3_INSTALL_DB_USE_EXISTING=0
+            - TYPO3_INSTALL_ADMIN_USER=admin
+            - TYPO3_INSTALL_ADMIN_PASSWORD=Password:joh316
+            - TYPO3_INSTALL_SITE_NAME=EXT:fal_s3 Dev Environment
+            - TYPO3_INSTALL_SITE_SETUP_TYPE=site
+            - TYPO3_INSTALL_WEB_SERVER_CONFIG=apache
+
+            # TYPO3 v13 config
+            - TYPO3_DB_DRIVER=mysqli
+            - TYPO3_DB_USERNAME=root
+            - TYPO3_DB_PASSWORD=root
+            - TYPO3_DB_HOST=db
+            - TYPO3_SETUP_ADMIN_EMAIL=admin@example.com
+            - TYPO3_SETUP_ADMIN_USERNAME=admin
+            - TYPO3_SETUP_ADMIN_PASSWORD=Password:joh316
+            - TYPO3_PROJECT_NAME=EXT:fal_s3 Dev Environment
+            - TYPO3_SERVER_TYPE=apache
+        volumes:
+            - type: bind
+              source: ../
+              target: /var/www/fal_s3
+              consistency: cached
+            - v12-data:/var/www/html/v12
+            - v13-data:/var/www/html/v13
+volumes:
+    v12-data:
+        name: "${DDEV_SITENAME}-v12-data"
+    v13-data:
+        name: "${DDEV_SITENAME}-v13-data"
