+++ .gitattributes
@@ -1,6 +1,8 @@
+Build export-ignore
 Documentation export-ignore
 Tests export-ignore
 .gitattributes export-ignore
 .gitignore export-ignore
 .travis.yml export-ignore
 README.md export-ignore
+.ddev export-ignore
