+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+- Support for TYPO3 13
+- DDEV Setup for development
+
+### Changed
+- Updated Documentation to the new documentation rendering standards
+
+### Removed
+- Support for TYPO3 11
+- Support for PHP 7 and 8.0
+
 ## [2.2.0] - 2024-10-25
 ### Added
 - Support for php 8.3
