+++ Documentation/Installation/Index.rst
@@ -1,0 +1,17 @@
+.. _installation:
+
+============
+Installation
+============
+
+As stated in the previous chapter, this extension must be installed using Composer to be able to use the AWS SDK.
+From that point on configuration is pretty straight-forward.
+
+Head over to your CLI and run the following command
+
+.. code-block:: bash
+
+    composer require maxserv/fal_s3
+
+This wil take care of installing all dependencies like the AWS SDK and this extension.
+After installing it's necessary to flush the system caches.
