--- Documentation/Settings.yml
@@ -1,21 +1,0 @@
-# This is the project specific Settings.yml file.
-# Place Sphinx specific build information here.
-# Settings given here will replace the settings of 'conf.py'.
-
----
-conf.py:
-  copyright: 2016
-  project: S3 driver for FAL
-  version: 0.0
-  release: 1.0.0
-  latex_documents:
-  - - Index
-    - fal_s3.tex
-    - S3 driver for FAL
-    - Arno Schoon
-    - manual
-  latex_elements:
-    papersize: a4paper
-    pointsize: 10pt
-    preamble: \usepackage{typo3}
-...
