+++ Documentation/Sitemap.rst
@@ -1,0 +1,7 @@
+:template: sitemap.html
+
+=======
+Sitemap
+=======
+
+.. The sitemap.html template will insert here the page tree automatically.
