+++ Documentation/guides.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<guides
+        xmlns="https://www.phpdoc.org/guides" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="https://www.phpdoc.org/guides ../vendor/phpdocumentor/guides-cli/resources/schema/guides.xsd"
+        links-are-relative="true"
+>
+    <extension
+            class="\T3Docs\Typo3DocsTheme\DependencyInjection\Typo3DocsThemeExtension"
+            project-home="https://extensions.typo3.org/extension/fal_s3"
+            project-contact="mailto:support@maxserv.com"
+            project-repository="https://github.com/MaxServ/t3ext-fal_s3"
+            project-issues="https://github.com/MaxServ/t3ext-fal_s3/issues"
+            edit-on-github-branch="master"
+            edit-on-github="MaxServ/t3ext-fal_s3"
+            typo3-core-preferred="stable"
+    />
+    <project
+            title="TYPO3 EXT:fal_s3"
+            release="3.0.0"
+            version="3.0"
+            copyright="since 2016 by MaxServ B.V."
+    />
+</guides>
