+++ composer.json
@@ -25,20 +25,27 @@
         "typo3-ter/fal_s3": "self.version"
     },
     "require": {
-        "php": ">=7.2, <8.4",
-        "aws/aws-sdk-php": "^3.324.5",
-        "typo3/cms-backend": "^10.4 || ^11.5 || ^12.4",
-        "typo3/cms-core": "^10.4 || ^11.5 || ^12.4",
-        "typo3/cms-extbase": "^10.4 || ^11.5 || ^12.4"
+        "php": "^8.1",
+        "aws/aws-sdk-php": "^3.338.0",
+        "typo3/cms-backend": "^12.4 || ^13.4",
+        "typo3/cms-core": "^12.4 || ^13.4",
+        "typo3/cms-extbase": "^12.4 || ^13.4"
     },
     "require-dev": {
-        "nimut/testing-framework": "^5.1 || ^6.0"
+        "typo3/testing-framework": "^8.2 || ^9.2"
     },
     "suggest": {
-        "causal/extractor": "^2.0"
+        "causal/extractor": "^2.5"
     },
     "config": {
-        "sort-packages": true
+        "allow-plugins": {
+            "typo3/class-alias-loader": true,
+            "typo3/cms-composer-installers": true
+        },
+        "sort-packages": true,
+        "bin-dir": ".Build/bin",
+        "lock": false,
+        "vendor-dir": ".Build/vendor"
     },
     "extra": {
         "typo3/cms": {
